Moved two classes from cmd2.py to utils.py.
