+++ cmd2/ansi.py
@@ -3,9 +3,9 @@
 Support for ANSI escape sequences which are used for things like applying style to text,
 setting the window title, and asynchronous alerts.
  """
+import enum
 import functools
 import re
-from enum import Enum
 from typing import IO, Any, List, Union
 
 import colorama
@@ -49,7 +49,7 @@ The default is ``STYLE_TERMINAL``.
 ANSI_STYLE_RE = re.compile(r'\x1b\[[^m]*m')
 
 
-class ColorBase(Enum):
+class ColorBase(enum.Enum):
     """
     Base class used for defining color enums. See fg and bg classes for examples.
