+++ docs/api/utils.rst
@@ -36,6 +36,33 @@ IO Handling
     :members:
 
 
+Tab Completion
+--------------
+
+.. autoclass:: cmd2.utils.CompletionMode
+
+    .. attribute:: NONE
+
+        Tab completion will be disabled during read_input() call. Use of custom
+        up-arrow history supported.
+
+    .. attribute:: COMMANDS
+
+        read_input() will tab complete cmd2 commands and their arguments.
+        cmd2's command line history will be used for up arrow if history is not
+        provided. Otherwise use of custom up-arrow history supported.
+
+    .. attribute:: CUSTOM
+
+        read_input() will tab complete based on one of its following parameters
+        (choices, choices_provider, completer, parser). Use of custom up-arrow
+        history supported
+
+.. autoclass:: cmd2.utils.CustomCompletionSettings
+
+    .. automethod:: __init__
+
+
 Text Alignment
 --------------
