+++ tests/test_completion.py
@@ -22,7 +22,6 @@ import pytest
 import cmd2
 from cmd2 import utils
 from examples.subcommands import SubcommandsExample
-
 from .conftest import complete_tester, normalize, run_cmd
 
 # List of strings used with completion functions
